requirements.txt
requests~=2.25.1
behave~=1.2.6
mysql-connector-python~=8.0.33
allure-behave
